// frontend/src/api/generated/schemas/answerResultDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * onBoard API
 * AI-powered interview preparation service
 * OpenAPI spec version: 1.0
 */
import type { AnswerNextQuestionDto } from "./answerNextQuestionDto";

export interface AnswerResultDto {
  answerId: string;
  score: number;
  feedback: string;
  isFullyClosed: boolean;
  recommendations: string[];
  currentOrder: number;
  status: string;
  isFinished: boolean;
  nextQuestion: AnswerNextQuestionDto | null;
}