// frontend/src/api/generated/schemas/sessionDetailDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * onBoard API
 * AI-powered interview preparation service
 * OpenAPI spec version: 1.0
 */
import type { SessionDetailDtoStatus } from "./sessionDetailDtoStatus";
import type { SessionQuestionDto } from "./sessionQuestionDto";
import type { SessionTechnologyLevelDto } from "./sessionTechnologyLevelDto";

export interface SessionDetailDto {
  id: string;
  userId: string;
  technologyLevelId: string;
  status: SessionDetailDtoStatus;
  /** @nullable */
  totalQuestions: number | null;
  currentOrder: number;
  /** @nullable */
  startedAt: string | null;
  /** @nullable */
  finishedAt: string | null;
  createdAt: string;
  technologyLevel: SessionTechnologyLevelDto;
  questions: SessionQuestionDto[];
}